fix BR 57371: distinguish between long and int for 64bit machines (however I doubt that this is the only place where kdevelop fails on 64bit - and I don't have a 64bit machine around for tests) CCMAIL: 57371-done@bugs.kde.org

+++ parts/trollproject/caret.h
@@ -16,8 +16,8 @@
 class Caret
 {
   public:
-  long      m_row;
-  long      m_idx;
+  int      m_row;
+  int      m_idx;
             Caret();
             Caret(const int row, const int idx);
             Caret(const Caret& value);
